// src/app/modules/components/table-exp/model/table.ts
export interface dataTable {
  archivado: null
  archivo: string;
  dependenciaCreador: string;
  fechaFin: string;
  fechaInicio: string;
  fechaProceso: string;
  idGdExpediente: number;
  idGdTrdDependencia: number;
  idGdTrdSerie: number;
  idGdTrdSubserie: number;
  nombreGdExpediente: string;
  numeroGdExpediente: string;
  serie: string;
  status: number;
  statusText: string;
  subserie: string;
  ubicacionGdExpediente: number;
  userCreador: string;
  userCreadorId: number;
}